+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.0.1 - 4 Mar 2024
+Bug fix:
+- incorrect array bracket limit
+- handle deep nested optional object property
+
 # 0.0.0 - 4 Mar 2024
 Feature:
 - initial release